+++ src/RabbitX/GlobalUsings.cs
@@ -14,6 +14,8 @@ global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.Logging;
 global using Microsoft.Extensions.Options;
 
+global using Microsoft.Extensions.Diagnostics.HealthChecks;
+
 global using Polly;
 
 global using RabbitMQ.Client;